requirements.txt
tensorflow
Pillow
json
matplotlib
numpy